Visualization changes
I delayed the binarification of gender so we keep all chr column as chr on visualization, only change them into 0 and 1 before the classification. I also removed the year column because it is no longer relevant for answering our questions. We will discuss the year range at the beginning so the reader know the data is still "recent", even though we no longer use it for our predictors. I also changed the visualization into a simpler one, comparing each predictors to the class, which is company.